pyproject.toml
[build-system]
requires = [
    "scikit-build-core",
    "pxr-arch-dev==25.5.1.*",
    "pxr-tf-dev==25.5.1.*",
    "pxr-js-dev==25.5.1.*",
    "pxr-trace-dev==25.5.1.*",
    "pxr-work-dev==25.5.1.*",
    "pxr-boost-dev==25.5.1.*",
    "pxr-tbb-dev==2022.1.0.*"
]
build-backend = "scikit_build_core.build"

[project]
name = "pxr-plug"
version = "25.5.1"
description = "Plugin Framework used in OpenUSD"
readme = "README.md"
requires-python = ">=3.8"
license-files = ["LICENSE.txt", "NOTICE.txt"]
keywords = ["pixar", "open-usd"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows :: Windows 10",
]
dependencies = [
    "pxr-arch==25.5.1.*",
    "pxr-tf==25.5.1.*",
    "pxr-js==25.5.1.*",
    "pxr-trace==25.5.1.*",
    "pxr-work==25.5.1.*",
    "pxr-boost==25.5.1.*",
    "pxr-tbb==2022.1.0.*"
]

[project.urls]
Repository = "https://github.com/untwine/pxr-plug"

[tool.scikit-build]
cmake.version = ">=3.21,<4"
cmake.args = [
    "-DBUILD_TESTS=OFF",
    "-DBUILD_SHARED_LIBS=ON",
    "-DBUILD_PYTHON_BINDINGS=ON",
    "-DCMAKE_INSTALL_LIBDIR=pxr/.libs",
    "-DCMAKE_INSTALL_INCLUDEDIR=pxr-plug/include",
    "-DCMAKE_INSTALL_DATAROOTDIR=pxr-plug",
    "-DCMAKE_INSTALL_PYTHON_LIBDIR=.",
]
wheel.exclude = ["pxr-plug"]
sdist.exclude = ["test", ".github"]
